osd/osd_types: fix pg_t::pool() return type (uint64_t -> int64_t)

+++ src/mon/OSDMonitor.cc
@@ -11490,7 +11490,7 @@ int OSDMonitor::_prepare_remove_pool(
   for (auto p = osdmap.pg_temp->begin();
        p != osdmap.pg_temp->end();
        ++p) {
-    if (p->first.pool() == (uint64_t)pool) {
+    if (p->first.pool() == pool) {
       dout(10) << __func__ << " " << pool << " removing obsolete pg_temp "
 	       << p->first << dendl;
       pending_inc.new_pg_temp[p->first].clear();
@@ -11500,7 +11500,7 @@ int OSDMonitor::_prepare_remove_pool(
   for (auto p = osdmap.primary_temp->begin();
       p != osdmap.primary_temp->end();
       ++p) {
-    if (p->first.pool() == (uint64_t)pool) {
+    if (p->first.pool() == pool) {
       dout(10) << __func__ << " " << pool
                << " removing obsolete primary_temp" << p->first << dendl;
       pending_inc.new_primary_temp[p->first] = -1;
@@ -11508,7 +11508,7 @@ int OSDMonitor::_prepare_remove_pool(
   }
   // remove any pg_upmap mappings for this pool
   for (auto& p : osdmap.pg_upmap) {
-    if (p.first.pool() == (uint64_t)pool) {
+    if (p.first.pool() == pool) {
       dout(10) << __func__ << " " << pool
                << " removing obsolete pg_upmap "
                << p.first << dendl;
@@ -11531,7 +11531,7 @@ int OSDMonitor::_prepare_remove_pool(
   }
   // remove any pg_upmap_items mappings for this pool
   for (auto& p : osdmap.pg_upmap_items) {
-    if (p.first.pool() == (uint64_t)pool) {
+    if (p.first.pool() == pool) {
       dout(10) << __func__ << " " << pool
                << " removing obsolete pg_upmap_items " << p.first
                << dendl;
